+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.5] - 2024-09-15
+### Cambios
+- Corrigiendo errores de movimientos en la pieza bishop, king y knihgt
+- Corriendo test de piezas de bishop, king y knight
+
 ## [0.2.4] - 2024-09-14
 ### Agregando
 - Implementando moetodo para testear la pieza pawn